% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageWithLabels.R
\name{colorscale}
\alias{colorscale}
\title{if you need an colorscale to you imagelables use this}
\usage{
colorscale(x, cex = 1, cex.axis = 0.5, col = heat.colors(12),
  digits = 2, zlim = NULL)
}
\arguments{
\item{cex}{cex}

\item{cex.axis}{cex.axis}

\item{col}{colors used}

\item{digits}{number of digits on color scale, default 2}

\item{zlim}{zlim}

\item{data}{the data matrix}
}
\description{
if you need an colorscale to you imagelables use this
}
\examples{

x = matrix(rnorm(20*30),ncol=20)
rownames(x) <- 1:30
colnames(x) <- letters[1:20]
colorscale2(x)
}

