% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CVlog}
\alias{CVlog}
\title{geometric coefficient of variation (CV for log transformed data)
Typically used to create and violin plot}
\usage{
CVlog(data, top = 30)
}
\arguments{
\item{data}{matrix}

\item{top}{remove top (default 30) CV}
}
\description{
geometric coefficient of variation (CV for log transformed data)
Typically used to create and violin plot
}
\examples{
dat <- matrix(rnorm(1000,10,5), ncol=20)
dim(dat)
cv <- CVlog(dat, top=5)
length(cv)
stopifnot(length(cv) == 45)
hist(cv)
}
\seealso{
CV
}

