% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2Matrix.R
\name{dat2Matrix}
\alias{dat2Matrix}
\title{converts sparse representation to dense where row and col can be character vectors}
\usage{
dat2Matrix(row, col, x)
}
\arguments{
\item{row}{row positions}

\item{col}{column positions}

\item{x}{values for matrix}
}
\value{
matrix
}
\description{
converts sparse representation to dense where row and col can be character vectors
}
