% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roc.R
\name{determineCut}
\alias{determineCut}
\title{determine best Accuracy cuttoff}
\usage{
determineCut(values, controls)
}
\arguments{
\item{values}{a factor of predicted classes}

\item{controls}{a factor of classes to be used as the true results}
}
\description{
determine best Accuracy cuttoff
}

