% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValuesForVolcano.R
\name{fisherExact}
\alias{fisherExact}
\title{get p-values using fishers exact test for count data}
\usage{
fisherExact(x, y, accessions)
}
\arguments{
\item{x}{- array}

\item{y}{- array}

\item{accessions}{accession string}
}
\value{
data frame with accessions, pval, pvaldj (BH adjusted p.values), fchange (log2 FC).
}
\description{
get p-values using fishers exact test for count data
}
\examples{
accessions <- letters
x <- sample(100,length(letters))
y <- sample(100,length(letters))
res <- fisherExact(x,y,accessions)
volcanoplot(res$fchange, res$pvaladj, labels = res$accessions)

}

