% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{getBlueWhiteRed}
\alias{getBlueWhiteRed}
\title{create blue white red palette}
\usage{
getBlueWhiteRed(length = 21)
}
\arguments{
\item{length}{nr of colors}
}
\description{
create blue white red palette
}
\examples{

library(scales)
show_col(getBlueWhiteRed(21))

}

