% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{getDiv}
\alias{getDiv}
\title{create divergent palette}
\usage{
getDiv(length = 10, pallete = 3)
}
\arguments{
\item{length}{nr of colors}
}
\description{
create divergent palette
}
\examples{

library(scales)
show_col(getDiv(21,4))

}

