% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{getGreensScale}
\alias{getGreensScale}
\title{create green color scale}
\usage{
getGreensScale(length = 10, increasing = FALSE)
}
\arguments{
\item{length}{nr of colors}

\item{increasing}{default FALSE}
}
\description{
create green color scale
}
\examples{

library(scales)
show_col(getGreensScale(5))

}

