% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{getRedScale}
\alias{getRedScale}
\title{create red color scale}
\usage{
getRedScale(length = 10)
}
\arguments{
\item{length}{nr of colors}
}
\description{
create red color scale
}
\examples{
library(scales)
show_col(getRedScale(12))
}

