% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getValuesForVolcano.R
\name{getValuesForVolcano}
\alias{getValuesForVolcano}
\title{running robust scaling of arefw}
\usage{
getValuesForVolcano(x, y)
}
\arguments{
\item{arefw}{- data array to scale}

\item{k}{- windows}

\item{scale}{- should also scaling be applied}
}
\value{
list with scaled data runmed used to center the data and runmad the running MAD used for scaling
}
\description{
running robust scaling of arefw
}
\examples{
a <- t(replicate(200,rnorm(20,runif(1,-3,3),1)))
b <- a[1:100,]
a <- a[101:200,]
boxplot(t(a[1:20,]))
boxplot(t(b[1:20,]))
res <- getValuesForVolcano(a,b)
volcanoplot(res$fchange , res$pval)
}

