% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageWithLabels.R
\name{imageColorscale}
\alias{imageColorscale}
\title{if you need an colorscale to you imagelables use this}
\usage{
imageColorscale(x, cex = 1, cex.axis = 0.5, col = heat.colors(12),
  digits = 2, zlim = NULL, breaks)
}
\arguments{
\item{x}{data the data matrix}

\item{cex}{cex}

\item{cex.axis}{cex.axis}

\item{col}{colors used}

\item{digits}{number of digits on color scale, default 2}

\item{zlim}{zlim}

\item{breaks}{optional argument passed to image (see image for more details)}
}
\description{
if you need an colorscale to you imagelables use this
}
\examples{
x = matrix(rnorm(20*30),ncol=20)
rownames(x) <- 1:30
colnames(x) <- letters[1:20]
imageColorscale(x)
}
