% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imageWithLabels.R
\name{imageWithLabels}
\alias{imageWithLabels}
\title{image plot with labels}
\usage{
imageWithLabels(x, col.labels = colnames(x), row.labels = rownames(x),
  cex = 1, cex.axis = 0.5, main = NULL, col = heat.colors(12),
  digits = 2)
}
\arguments{
\item{x}{matrix}

\item{col.labels}{rownames(x)}

\item{row.labels}{colnames(x)}

\item{cex}{size of labels}

\item{cex.axis}{size of axis lables}

\item{main}{main title}

\item{col}{color map for matrix}

\item{digits}{number of digits on colorscale, default 2}
}
\description{
image plot with labels
}
\examples{
x = matrix(rnorm(20*30),ncol=20)
rownames(x) <- 1:30
colnames(x) <- letters[1:20]
imageWithLabels(x)
}

