% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{package}
\name{imsbInfer}
\alias{imsbInfer}
\alias{imsbInfer-package}
\title{A package with some utility functions for the analysis of swath data}
\description{
A package with some utility functions for the analysis of swath data
}

