% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roc.R
\name{makeROCplot}
\alias{makeROCplot}
\title{create density and ROC plot}
\usage{
makeROCplot(cases, controls, label = "", xlim = c(-5, 5), abline = 0)
}
\arguments{
\item{cases}{a factor of predicted classes}

\item{controls}{a factor of classes to be used as the true results}
}
\description{
create density and ROC plot
}
\examples{
cases <- rnorm(200,-1,3)
controls <- rnorm(300,2,4)
}

