% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_tibble.R
\name{matrix_to_tibble}
\alias{matrix_to_tibble}
\title{Matrix to tibble (taken from tidyquant)}
\usage{
matrix_to_tibble(x, preserve_row_names = TRUE, ...)
}
\arguments{
\item{x}{a matrix}

\item{preserve_row_names}{should rownames be kept (default TRUE)}

\item{...}{further parameters passed to as_tibble}
}
\description{
Matrix to tibble (taken from tidyquant)
}
