% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigroupVolcano.R
\name{multigroupVolcano}
\alias{multigroupVolcano}
\title{plot volcano given multiple conditions}
\usage{
multigroupVolcano(misspX, colour, effect = "fc", type = "p.adjust",
  Condition = "condition", xintercept = c(-2, 2), pvalue = 0.05,
  label = "row.names")
}
\arguments{
\item{misspX}{data in long format}

\item{effect}{column containing effect sizes}

\item{type}{column containing p-values, q.values etc}

\item{Condition}{column with condition}

\item{xintercept}{fc thresholds}

\item{pvalue}{pvalue threshold}

\item{label}{column containing labels}
}
\description{
plot volcano given multiple conditions
}
