% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSampleSingleSidedVolcano.R
\name{onesamplegreaterT}
\alias{onesamplegreaterT}
\title{One sample single sided t-test on matrix}
\usage{
onesamplegreaterT(ddtas, adjust = TRUE)
}
\arguments{
\item{ddtas}{- matrix}

\item{adjust}{- should p-values be adjusted}
}
\description{
One sample single sided t-test on matrix
}
\examples{

data <- matrix(rnorm(100,1,1), ncol=5)
dim(data)
rownames(data) <- 1:20
onesamplegreaterT(data)

}
