% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypairs.R
\name{panel.cor}
\alias{panel.cor}
\title{correlation panel for pairs plot function (used as default in mypairsSmooth)}
\usage{
panel.cor(x, y, digits = 2, cex.cor, ...)
}
\arguments{
\item{x}{numeric data}

\item{y}{numeric data}
}
\description{
correlation panel for pairs plot function (used as default in mypairsSmooth)
}
