% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgenesisInput.R
\name{ProgenesisBuildAnnotation}
\alias{ProgenesisBuildAnnotation}
\title{build annotation from column names}
\usage{
ProgenesisBuildAnnotation(data)
}
\arguments{
\item{data}{tibble returned by ProgenesisRead}
}
\value{
list of tibbles data -
}
\description{
build annotation from column names
}
\examples{

file = file.path(path.package("quantable"),"extdata/PG/ProteinMeasurement_inclSingleHits_hi3.csv" )
tmp <- ProgenesisRead(file)
colnames(tmp)
xx <- ProgenesisBuildAnnotation(tmp)
head(xx$anno)
colnames(xx$data)
colnames(tmp)
}
