% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgenesisInput.R
\name{ProgenesisRead}
\alias{ProgenesisRead}
\title{reads file exportet from progenesis
ProgenesisRead}
\usage{
ProgenesisRead(file, sep = ",")
}
\arguments{
\item{file}{path to progenesis peptide or protein file}

\item{sep}{separator used (progenesis uses depending language settings a , or ;)}
}
\description{
reads file exportet from progenesis
ProgenesisRead
}
\examples{
file = file.path(path.package("quantable"),"extdata/PG/PeptideMeasurement_inclSingleHits_hi3.csv" )
tmp <- ProgenesisRead(file)
colnames(tmp)
head(tmp)

file = file.path(path.package("quantable"),"extdata/PG/ProteinMeasurement_inclSingleHits_hi3.csv" )
tmp <- ProgenesisRead(file)
colnames(tmp)
}
