% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{package}
\name{quantable}
\alias{quantable}
\alias{quantable-package}
\title{Streamline descriptive analysis of quantitative data matrices}
\description{
Streamline descriptive analysis of quantitative data matrices
}

