% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDecorellated.R
\name{removeDecorrelated}
\alias{removeDecorrelated}
\title{remove decorrelated rows}
\usage{
removeDecorrelated(ff, corThreshold = 0.7, tr = identity)
}
\arguments{
\item{ff}{matrix or data frame}

\item{corThreshold}{correlation threshold}

\item{tr}{intensity transformation to apply}
}
\description{
remove decorrelated rows
}
