% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNArows.R
\name{removeNArows}
\alias{removeNArows}
\title{Removes rows with more than thresh NA's from matrix}
\usage{
removeNArows(obj, thresh = 0)
}
\arguments{
\item{obj}{matrix or dataframe}

\item{thresh}{- maximum number of NA's / row - if more the row will be removed}
}
\value{
matrix
}
\description{
Removes rows with more than thresh NA's from matrix
}
\examples{

x = matrix(rnorm(10*10),ncol=10)
dim(x)
x[3,3] = NA
x = removeNArows(x)
dim(x)
}
