% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustscale.R
\name{robustscale}
\alias{robustscale}
\title{robust scaling
uses median an mad instead of mean and row
applies the scaling to the columns (samples) by default}
\usage{
robustscale(data, dim = 2, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{matrix or data.frame}

\item{dim}{- should rows (1) or columns (2:default) be scaled}

\item{center}{- subract median (default:TRUE)}

\item{scale}{- scale by mad  (default:FALSE)}
}
\description{
robust scaling
uses median an mad instead of mean and row
applies the scaling to the columns (samples) by default
}
\examples{
tmp = matrix(rep((1:100),times = 4) + rnorm(100*4,0,3),ncol=4)
mean = c(20,30,10,40)
sd = c(4,3,4,5)
tmp = sweep(tmp,2,sd,"*")
tmp = sweep(tmp,2,mean,"+")
boxplot(tmp)
tmp = robustscale(tmp)
boxplot(tmp$data)
}

