% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runrobscale.R
\name{runTICscale}
\alias{runTICscale}
\title{running total ion count scaling (TIC)}
\usage{
runTICscale(arefw, k = 101)
}
\arguments{
\item{arefw}{a series to scale}

\item{k}{- the smoothing window}
}
\value{
list with fields scaled - contains scaled data and mean - averages of window k
}
\description{
running total ion count scaling (TIC)
}
\examples{
res = c(rnorm(1000,3,2),rnorm(2000,8,1))
res2 = runTICscale(res)
plot(res,type="p",pch=".",col=1,cex=0.5)
lines(1:length(res),res2$mean,col=3)
points(res2$scaled, pch=".",cex=3,col=2)

}
\seealso{
correctIntRTv2 for context
}
