% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleByGroup.R
\name{scaleByGroup}
\alias{scaleByGroup}
\title{scale data given group}
\usage{
scaleByGroup(data, protGroup, plot = FALSE)
}
\arguments{
\item{data}{matrix}

\item{protGroup}{names matching rownames of data}

\item{plot}{plot diagnostics}
}
\description{
scale data given group
}

