% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simpleheatmap.R
\name{simpleheatmap}
\alias{simpleheatmap}
\title{heatmap2 facade}
\usage{
simpleheatmap(pln, main = "", distf = dist, hclustf = hclust,
  labRow = "", palette = div_gradient_pal(low = "blue", mid = "white", high
  = "red")(seq(0, 1, length = 21)), ...)
}
\arguments{
\item{pln}{or dataframe with numerical values}

\item{main}{title}

\item{distf}{distance function}

\item{hclustf}{clustering function}

\item{labRow}{row labels}

\item{palette}{color palette}

\item{...}{other parameters to heatmap.2}
}
\description{
heatmap2 facade
}
\examples{
data(SDat)
par(mar=c(5,5,5,5))
simpleheatmap(SDat$Intensity,ColSideColors=c("red","blue","pink"))
simpleheatmap(SDat$Intensity)
}

