% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleheatmap3.R
\name{simpleheatmap3}
\alias{simpleheatmap3}
\title{heatmap3 facade}
\usage{
simpleheatmap3(pln, main = "", distf = dist, hclustf = hclust,
  labRow = "", labCol = "", palette = getBlueWhiteRed(),
  margins = c(5, 5), scale = "none", plot = TRUE,
  nrOfClustersCol = 3, nrOfClustersRow = 3, ...)
}
\arguments{
\item{pln}{matrix or dataframe with numerical values}

\item{main}{title}

\item{distf}{distance function}

\item{hclustf}{clustering function}

\item{labRow}{row labels}

\item{labCol}{column labels}

\item{palette}{color palette}

\item{margins}{control margins of heatmap}

\item{scale}{c(row, column or none)}

\item{plot}{logical; If TRUE simpleheatmap3 returns modified heatmap3 object, if FALSE returns a data.frame with clusterIDs. Default is TRUE}

\item{nrOfClustersCol}{number of distinct clusters for cutree function on columns. Default is 3}

\item{nrOfClustersRow}{number of distinct clusters for cutree function on rows. Default is 3}

\item{...}{other parameters to heatmap3}
}
\description{
heatmap3 facade
}
\examples{
df <- matrix(rnorm(150), ncol = 10) # 15x10 matrix
clust <- simpleheatmap3(pln = df,
 main = "",
 distf = dist,
 hclustf = hclust,
 labRow = "",
 plot = FALSE,
 nrOfClustersRow = 3) 
# plot = F will return a list containing two data.frames 
# with the specified number of distinct clusters in nrOfClustersRow
# and nrOfClustersCol
hmp3 <- simpleheatmap3(pln = df,
 main = "",
 distf = dist,
 hclustf = hclust,
 labRow = "",
 plot = FALSE,
 nrOfClustersRow = 3)  
}
