% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumtop.R
\name{sumtop}
\alias{sumtop}
\title{sums top rows of matrix
uses median of row to determine row order}
\usage{
sumtop(x, top = 3)
}
\arguments{
\item{x}{a matrix}

\item{top}{how many to rows to sum (default 3)}
}
\description{
sums top rows of matrix
uses median of row to determine row order
}
\examples{
tmp <- rbind(rep(1,times = 4), rep(2,times=4), rep(3,times=4), rep(4,times=4))
res <- sumtop(tmp)
stopifnot(res == c(9,9,9,9))

}
