% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot.R
\name{unpivot}
\alias{unpivot}
\title{unpivot data matrix}
\usage{
unpivot(m)
}
\arguments{
\item{m}{matrix or dataframe}
}
\description{
unpivot data matrix
}
\examples{
x = matrix(1:25,ncol=5)
x = as.data.frame(x)
colnames(x) = letters[1:5]
rownames(x) = LETTERS[1:5]
unpivot(x)
}
