% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uppertriang.R
\name{uppertriang}
\alias{uppertriang}
\title{get values of upper triangle from matrix}
\usage{
uppertriang(mat)
}
\arguments{
\item{mat}{matrix}
}
\description{
get values of upper triangle from matrix
}
\examples{
t = matrix(1:25,ncol=5)
uppertriang(t)
}

