% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano2G.R
\name{volcano2G}
\alias{volcano2G}
\title{Volcano plot using ggplot and ggrepel}
\usage{
volcano2G(foldchange, pvals, labels, pthresh = 0.1, log2FCThresh = 0.5,
  main = NULL)
}
\arguments{
\item{foldchange}{vector with fold changes}

\item{pvals}{vector with pvalues}

\item{labels}{vector with labels}

\item{pthresh}{pvalue threshold}

\item{log2FCThresh}{log2 FC threshold}

\item{main}{main title}
}
\description{
Volcano plot using ggplot and ggrepel
}
\examples{
library(quantable)
foldchange <- rnorm(1000)
pvals <-rexp(1000)
volcano2G(foldchange, pvals,labels=rep("abcde", length(pvals)),
  pthresh=0.1, log2FCThresh=0.5,main='test')

}

