% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{volcano plot}
\usage{
volcanoplot(foldchange, pvals, pthresh = 0.05, ratiothresh = 2,
  xlab = "log2(T/N)", ylab = "-log10(P)", labels = NULL, cex = 0.6,
  cex.point = 1, xlim = NULL, main = NULL)
}
\arguments{
\item{foldchange}{- fold change values}

\item{pvals}{pvalues}

\item{pthresh}{pvalue threshold}

\item{ratiothresh}{threshold of foldchange}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{labels}{- optional labels}

\item{cex}{size of labels}

\item{cex.point}{- point size}

\item{xlim}{- xlim}

\item{main}{- main title}
}
\description{
volcano plot
}
\examples{
foldchange <- rnorm(1000)
pval <-rexp(1000)
abline(v=0.05,col=2)

volcanoplot(foldchange, pval,pthresh=0.1, ratiothresh=1,main='test')

}

