% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.vector.R
\name{write.tab}
\alias{write.tab}
\title{write table in tab delimited no quotes no row.names
(usefull for exporting i.e. list of foldchanges)
uses write.table 
TODO : validate if not redundant with other write functions}
\usage{
write.tab(x, file)
}
\arguments{
\item{x}{vector to write}

\item{file}{file to write to}
}
\description{
write table in tab delimited no quotes no row.names
(usefull for exporting i.e. list of foldchanges)
uses write.table 
TODO : validate if not redundant with other write functions
}
