% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_replace_xxx.R
\name{xxx_replace_xxx}
\alias{xxx_replace_xxx}
\title{replace patterns (vector) with replacements (vector) in string or string vector.
uses gsub and perl=TRUE
taken from \url{https://stackoverflow.com/questions/26676045}}
\usage{
xxx_replace_xxx(string, patterns, replacements)
}
\arguments{
\item{string}{string or vector of strings to replace in}

\item{patterns}{pattern or vector of patterns to replace}

\item{replacements}{replacements string, needs to have same length as replacment}
}
\description{
replace patterns (vector) with replacements (vector) in string or string vector.
uses gsub and perl=TRUE
taken from \url{https://stackoverflow.com/questions/26676045}
}
