% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/altmanbland.R
\name{altmanbland}
\alias{altmanbland}
\title{altman-bland plot}
\usage{
altmanbland(x, y, main = "", pch = ".", log = "")
}
\arguments{
\item{x}{- input intensities}

\item{y}{- input intensities}

\item{main}{plotting parameters}

\item{pch}{- plot character}

\item{log}{- should the x y axis be log transformed possible values "x" "y" or "xy"}
}
\description{
plots abs(x-y) against (y+x)/2
}
\examples{
x <- seq(1:300)/10
x <- x + rnorm(length(x),0,0.5)
y <- seq(1:300)/10
y <- y + rnorm(length(y),0,0.5)
altmanbland(y,x)
}

