% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imageWithLabels.R
\name{colorscale}
\alias{colorscale}
\title{if you need an colorscale to you imagelables use this}
\usage{
colorscale(data, colors = heat.colors(12))
}
\arguments{
\item{data}{the data matrix}

\item{colors}{used}
}
\description{
if you need an colorscale to you imagelables use this
}

