% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roc.R
\name{determineCut}
\alias{determineCut}
\title{determine best Accuracy cuttoff}
\usage{
determineCut(cases, controls, plot = FALSE, scanstep = 0.01)
}
\arguments{
\item{cases}{a factor of predicted classes}

\item{controls}{a factor of classes to be used as the true results}

\item{plot}{create plot of threasholds versus Accuracy}

\item{scanstep}{step size for threshold estimation}
}
\description{
determine best Accuracy cuttoff
}
\examples{
library(pROC)
cases <- rnorm(100,-1,1.1)
controls <- rnorm(500,1,1.1)
cut <- determineCut(cases,controls)
makeROCplot(cases,controls,abline=cut)
}

