% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{getBlueScale}
\alias{getBlueScale}
\title{create blue color scale}
\usage{
getBlueScale(length = 10, increasing = FALSE)
}
\arguments{
\item{length}{nr of colors}

\item{increasing}{default FALSE}
}
\description{
create blue color scale
}
\examples{

library(scales)
show_col(getBlueScale(5))

}
