% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{getDiv}
\alias{getDiv}
\title{create divergent palette}
\usage{
getDiv(length = 11, pallete = 3, increasing = TRUE)
}
\arguments{
\item{length}{nr of colors}

\item{pallete}{there are a few divergent palletes in brewer_pal, default 1}

\item{increasing}{default TRUE}
}
\description{
create divergent palette
}
\examples{

library(scales)
show_col(getDiv(21,4))
show_col(getDiv(21,4, increasing=FALSE))
}

