% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{getRedScale}
\alias{getRedScale}
\title{create red color scale}
\usage{
getRedScale(length = 10, increasing = FALSE)
}
\arguments{
\item{length}{nr of colors}

\item{increasing}{default FALSE}
}
\description{
create red color scale
}
\examples{

library(scales)
show_col(getRedScale(12))

}

