% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imageWithLabels.R
\name{imageWithLabels}
\alias{imageWithLabels}
\title{image plot with labels}
\usage{
imageWithLabels(x, labels = colnames(x), cex = 1, cex.axis = 0.5,
  main = NULL, col = heat.colors(12))
}
\arguments{
\item{x}{- matrix}

\item{labels}{- colnames(x)}

\item{cex}{- size of labels}

\item{cex.axis}{- size of axis lables}

\item{main}{- main title}

\item{col}{- color map for matrix}
}
\description{
image plot with labels
}
\examples{
x = matrix(rnorm(20*20),ncol=20)
imageWithLabels(x)
}

