% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roc.R
\name{makeROCplot}
\alias{makeROCplot}
\title{create density and ROC plot}
\usage{
makeROCplot(cases, controls, label = "", xlim = NULL, abline = NULL)
}
\arguments{
\item{cases}{a factor of predicted classes}

\item{controls}{a factor of classes to be used as the true results}

\item{label}{image main label}

\item{xlim}{x axis extend}

\item{abline}{draws vertical abline to indicate best threshold}
}
\description{
create density and ROC plot
}
\examples{
library(pROC)
cases <- rnorm(100,-1,1.5)
controls <- rnorm(300,1,1.5)
makeROCplot(cases,controls)
}

