% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mypairs.R
\name{mypairs}
\alias{mypairs}
\title{normal pairs plot with different pch and plus abline}
\usage{
mypairs(dataframe, ...)
}
\arguments{
\item{dataframe}{data matrix or data.frame as normally passed to pairs}

\item{...}{params usually passed to pairs}
}
\description{
normal pairs plot with different pch and plus abline
}
\examples{
data(SDat)
mypairs(SDat$Intensity,log="xy",main="small data")
}
\seealso{
also \code{\link{pairs}}
}

