% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/robustscale.R
\name{robustscale}
\alias{robustscale}
\title{robust scaling
uses median an mad instead of mean and row
applies the scaling to the columns (samples) by default}
\usage{
robustscale(data, dim = 2, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{matrix or data.frame}

\item{dim}{- should rows (1) or columns (2:default) be scaled}

\item{center}{- subract median (default:TRUE)}

\item{scale}{- scale by mad  (default:FALSE)}
}
\description{
robust scaling
uses median an mad instead of mean and row
applies the scaling to the columns (samples) by default
}
\examples{
data(SDat)
boxplot(SDat$Intensity)
tmp = asinh(SDat$Intensity)
tmp = robustscale(tmp)
boxplot(tmp$data)
}

