% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleByGroup.R
\name{scaleByGroup}
\alias{scaleByGroup}
\title{scale data given group}
\usage{
scaleByGroup(data, protGroup, plot = FALSE, scale = TRUE, center = TRUE)
}
\arguments{
\item{data}{matrix}

\item{protGroup}{names matching rownames of data}

\item{plot}{plot diagnostics}

\item{scale}{scale default TRUE}

\item{center}{center default TRUE}
}
\description{
scale data given group
}
