% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleheatmap.R
\name{simpleheatmap}
\alias{simpleheatmap}
\title{heatmap2 facade}
\usage{
simpleheatmap(pln, main = "", distf = dist, hclustf = hclust,
  labRow = "", palette = getBlueWhiteRed(), margins = c(5, 5),
  scale = "none", ...)
}
\arguments{
\item{pln}{or dataframe with numerical values}

\item{main}{title}

\item{distf}{distance function}

\item{hclustf}{clustering function}

\item{labRow}{row labels}

\item{palette}{color palette}

\item{margins}{control margins of heatmap}

\item{scale}{c(row, column or none)}

\item{...}{other parameters to heatmap.2}
}
\description{
heatmap2 facade
}
\examples{
tmp = matrix(rep((1:100),times = 4) + rnorm(100*4,0,3),ncol=4)
mean = c(20,30,10,40)
sd = c(4,3,4,5)
tmp = sweep(tmp,2,sd,"*")
tmp = sweep(tmp,2,mean,"+")
par(mar=c(5,5,5,5))
simpleheatmap(tmp,ColSideColors=c("red","blue","pink","black"))
simpleheatmap(tmp)
}
