% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleheatmap3.R
\name{simpleheatmap3}
\alias{simpleheatmap3}
\title{heatmap3 facade}
\usage{
simpleheatmap3(pln, main = "", distf = dist, hclustf = hclust,
  labRow = "", labCol = "", palette = getBlueWhiteRed(),
  margins = c(5, 5), scale = "none", plot = TRUE,
  nrOfClustersCol = 3, nrOfClustersRow = 3,
  suppressColSideCols = FALSE, ColSideLabs = "", RowSideLabs = "",
  ...)
}
\arguments{
\item{pln}{matrix or dataframe with numerical values}

\item{main}{title}

\item{distf}{distance function}

\item{hclustf}{clustering function}

\item{labRow}{row labels}

\item{labCol}{column labels}

\item{palette}{color palette}

\item{margins}{control margins of heatmap}

\item{scale}{c(row, column or none)}

\item{plot}{logical; If TRUE simpleheatmap3 returns modified heatmap3 object, if FALSE returns a data.frame with clusterIDs. Default is TRUE}

\item{nrOfClustersCol}{number of distinct clusters for cutree function on columns. Default is 3}

\item{nrOfClustersRow}{number of distinct clusters for cutree function on rows. Default is 3}

\item{suppressColSideCols}{Option to suppress color side labelling per default}

\item{...}{other parameters to heatmap3}
}
\description{
heatmap3 facade
}
\examples{
df <- matrix(rnorm(150), ncol = 10) # 15x10 matrix without NAs

clust <- simpleheatmap3(pln = df,
                        main = "",
                        distf = dist,
                        hclustf = hclust,
                        labRow = "",
                        plot = FALSE,
                        nrOfClustersRow = 3,
                        suppressColSideCols = TRUE) 
# plot = F will return a list containing two data.frames 
# with the specified number of distinct clusters in nrOfClustersRow
# and nrOfClustersCol
clust
hmp3 <- simpleheatmap3(pln = df,
                       main = "",
                       distf = dist,
                       hclustf = hclust,
                       labRow = "",
                       plot = TRUE,
                       nrOfClustersRow = 3, 
                       nrOfClustersCol = 3, 
                       suppressColSideCols = FALSE)
# plot = TRUE will result in the same list and additionally plot the heatmap. 
# suppressColSideCols = FALSE will leave the heatmap with ColSideColors

stopifnot(all(clust$Row$clusterID==hmp3$Row$clusterID))

df[3,1:7] <- NA # Seed some missing values in line 3

hmp3_withNAs <- simpleheatmap3(pln = df,
                       main = "",
                       distf = dist,
                       hclustf = hclust,
                       labRow = paste0("ABC", 1:nrow(df)),
                       plot = TRUE,
                       nrOfClustersRow = 3, 
                       nrOfClustersCol = 3, 
                       suppressColSideCols = TRUE)
                       
# Since line 3 contained more than ncol(df)/2 missing values 
# it is removed from the clustering.

stopifnot(! 3 \%in\% hmp3_withNAs$Row$rowID)
}
