% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleheatmap3.R
\name{simpleheatmap3}
\alias{simpleheatmap3}
\title{heatmap3 facade}
\usage{
simpleheatmap3(pln, main = "", distf = dist, hclustf = hclust,
  labRow = "", palette = getBlueWhiteRed(), margins = c(5, 5),
  scale = "none", plot = TRUE, nrOfClusters = 3, ...)
}
\arguments{
\item{pln}{matrix or dataframe with numerical values}

\item{main}{title}

\item{distf}{distance function}

\item{hclustf}{clustering function}

\item{labRow}{row labels}

\item{palette}{color palette}

\item{margins}{control margins of heatmap}

\item{scale}{c(row, column or none)}

\item{plot}{logical; If TRUE simpleheatmap3 returns modified heatmap3 object, if FALSE returns a data.frame with clusterIDs. Default is TRUE}

\item{nrOfClusters}{number of distinct clusters for cutree function. Default is 3}

\item{...}{other parameters to heatmap3}
}
\description{
heatmap3 facade
}
\examples{
df <- matrix(rnorm(150), ncol = 10) # 15x10 matrix
clust <- simpleheatmap3(pln = df, main = "", distf = dist, hclustf = hclust, labRow = "", plot = F, nrOfClusters = 3) # plot = F will return a data.frame
# with the specified number of distinct clusters in nrOfClusters
hmp3 <- simpleheatmap3(pln = df, main = "", distf = dist, hclustf = hclust, labRow = "", plot = F, nrOfClusters = 3)  plot = T will plot the heatmap with row and column dendrogram
}
