% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcano2G}
\alias{volcano2G}
\title{Volcano plot using ggplot and ggrepel}
\usage{
volcano2G(foldchange, pvals, labels, pthresh = 0.1, foldchangethresh = 0.5,
  main = "test")
}
\description{
Volcano plot using ggplot and ggrepel
}
\examples{
library(quantable)
foldchange <- rnorm(1000)
pvals <-rexp(1000)
volcano2G(foldchange, pvals,labels=rep("abcde", length(pvals)), pthresh=0.1, foldchangethresh=0.5,main='test')

}

