% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{volcano plot}
\usage{
volcanoplot(foldchange, pvals, pthresh = 0.05, foldchangethresh = 1,
  xlab = "log2(T/N)", ylab = "-log10(P)", labels = NULL, cex = 0.6,
  cex.point = 1, xlim = NULL, main = NULL, biasAdjust = FALSE)
}
\arguments{
\item{foldchange}{- fold change values}

\item{pvals}{pvalues}

\item{pthresh}{pvalue threshold}

\item{foldchangethresh}{threshold of foldchange}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{labels}{- optional labels}

\item{cex}{size of labels}

\item{cex.point}{- point size}

\item{xlim}{- xlim}

\item{main}{- main title}

\item{biasAdjust}{- if bias in foldchanges exists (i.e. if median of fold changes does not equal 0) you can use this option to adjust for it.}
}
\description{
volcano plot
}
\examples{

library(quantable)
foldchange <- rnorm(1000)
pvals <-rexp(1000)
length(foldchange)
length(pvals)
volcanoplot(foldchange, pvals, pthresh=0.1, foldchangethresh=1,main='test')
volcanoplot(foldchange, pvals, pthresh=0.1,
 foldchangethresh=1,main='test', labels=rep("abcde", length(pvals)))

volcanoplot(foldchange, pvals,pthresh=0.1, foldchangethresh=3,main='test')
abline(v=0.05,col=2)

}
