% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.vector.R
\name{write.vector}
\alias{write.vector}
\title{write vectors as single column table
(usefull for exporting i.e. protein id's)
uses write.table}
\usage{
write.vector(x, file)
}
\arguments{
\item{x}{vector to write}

\item{file}{file to write to}
}
\description{
write vectors as single column table
(usefull for exporting i.e. protein id's)
uses write.table
}

